Underground Polygon Area resource Calculation and Cost Analysis with C
Utilizes OpenGL and CURL libraries to simulate the exploration and economic evaluation of underground resources based on specific coordinate data. The project reads shapes composed of certain coordinate points, calculates the areas of these shapes, and analyzes the platform and drilling requirements for each shape. Additionally, it conducts a cost analysis by considering specified prices and unit values, calculating the cost incurred against the total reserves. The results are visualized through OpenGL and presented to the user.